requirements.txt
appium_python_client == 2.2.0
sauceclient >= 0.2.1
pytest==7.1.2
pytest-xdist
pytest-randomly